added factories, refactorings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@9148 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 07575d3f773adf5eeeb72b9182a943b0cb7b7b75

+++ core/scheduler.slamodel/src/de/uniol/inf/is/odysseus/scheduler/slamodel/Penalty.java
@@ -11,6 +11,6 @@ public interface Penalty {
 	/**
 	 * @return the amount that must be payed if a penalty is payable
 	 */
-	public int getCost();
+	public double getCost();
 	
 }
